+++ CHANGELOG.rst
@@ -6,6 +6,8 @@ Changelog
 Next
 ----
 
+- Breaking change: Use ``:substitutions:`` option on ``code-block`` or ``prompt`` rather than new directives.
+
 2020.05.23.0
 ------------